add videoLink

+++ solarboat/src/main/java/hu/schdesign/solarboat/dao/NewsRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NewsRepository extends CrudRepository<News, Integer> {
+public interface NewsRepository extends CrudRepository<News, Long> {
+
 }
